requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
anthropic==0.8.1
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
loguru==0.7.2